requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.5
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
pip==21.2.2
psycopg2==2.8.6
pytz==2021.3
setuptools==58.0.4
sqlparse==0.4.1
typing-extensions==3.10.0.2
wheel==0.37.1
wincertstore==0.2
zipp==3.7.0
